fix: changed the structure and moved the date
