--- src/assets/social/social.js
@@ -1,46 +1,0 @@
-import { FiSend } from 'react-icons/fi';
-import { ImWhatsapp } from 'react-icons/im';
-import { FiFacebook } from 'react-icons/fi';
-import { FiInstagram } from 'react-icons/fi';
-
-export const fullSocial = [
-  {
-    id: 1,
-    name: 'Link to telegram',
-    component: FiSend,
-    href: 'https://web.telegram.org/z/',
-  },
-  {
-    id: 2,
-    name: 'link to whatsapp',
-    component: ImWhatsapp,
-    href: 'https://www.whatsapp.com/?lang=ru',
-  },
-  {
-    id: 3,
-    name: 'Link to facebook',
-    component: FiFacebook,
-    href: 'https://www.facebook.com/',
-  },
-  {
-    id: 4,
-    name: 'Link to instagram',
-    component: FiInstagram,
-    href: 'https://www.instagram.com/',
-  },
-];
-
-export const social = [
-  {
-    id: 1,
-    name: 'Link to telegram',
-    component: FiSend,
-    href: 'https://web.telegram.org/z/',
-  },
-  {
-    id: 2,
-    name: 'Link to whatsapp',
-    component: ImWhatsapp,
-    href: 'https://www.whatsapp.com/?lang=ru',
-  },
-];
